chore: change package name to io.d11.aerospike
