--- src/main/java/com/dream11/aerospike/package-info.java
@@ -1,4 +1,0 @@
-@ModuleGen(groupPackage = "com.dream11.aerospike", name = "vertx-aerospike-client")
-package com.dream11.aerospike;
-
-import io.vertx.codegen.annotations.ModuleGen;
