+++ src/main/java/io/d11/aerospike/package-info.java
@@ -1,0 +1,4 @@
+@ModuleGen(groupPackage = "io.d11", name = "vertx-aerospike-client")
+package io.d11.aerospike;
+
+import io.vertx.codegen.annotations.ModuleGen;
